+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [$component]: When accessing functions which use $refs or other magic properties, `this` was not bound correctly.
 
 ## [0.5.1] - 2020-11-30
 ### Fixed