+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [`$undo/$track/$history`]: Adds a set of helpers to track component state and revert changes on demand
+- [`$refresh`]: Adds a helper to refresh components
 ## [0.5.1] - 2020-11-30
 ### Fixed
 - [$scroll]: When using css selectors or Alpine reference, element position to scroll to was some times a decimal number (e.g. rem values could resolve to a decimal number) and $scroll was triggering an error.