+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [`$component`]: Added support for accessing magic properties/helper via $component/$parent ($parent.$parent.foo is now a thing!!!)
+
+### Changed
+- [`$component`]: $component and $parent are now deferred for a few ms if the observed component is not ready
 
 ## [0.6.0] - 2021-02-05
 ### Fixed