+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0] - 2021-02-13
 ### Added
 - [`$component`]: Added support for accessing magic properties/helper via $component/$parent ($parent.$parent.foo is now a thing!!!)